@@ -6,9 +6,10 @@
     "clean": "npm run clean:lib && npm run clean:coverage",
     "clean:lib": "rimraf ./lib",
     "clean:coverage": "rimraf ./coverage",
+    "flow": "flow",
     "lint": "npm run lint:es",
-    "lint:es": "eslint --config .eslintrc src/**/*.js",
-    "prepublish": "babel -d lib ./src/ && rm lib/**/*.test.js",
+    "lint:es": "eslint --config .eslintrc src/**",
+    "prepublish": "babel -d lib ./src/ && rimraf lib/**/*.test.js",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post",
     "test": "jest --no-cache",
     "test:watch": "jest --watch",
@@ -17,7 +18,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/acateland/react-flow-designer.git"
+    "url": "https://github.com/Talend/react-flow-designer.git"
   },
   "author": "Axel Cateland <acateland@talend.com>",
   "license": "Apache-2.0",
@@ -34,34 +35,37 @@
     ]
   },
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.16.0",
-    "babel-eslint": "7.0.0",
-    "babel-jest": "15.0.0",
-    "babel-loader": "6.2.5",
+    "babel-cli": "^6.22.2",
+    "babel-core": "^6.22.1",
+    "babel-eslint": "7.1.1",
+    "babel-jest": "^18.0.0",
+    "babel-loader": "^6.2.10",
     "babel-plugin-add-module-exports": "0.2.1",
-    "babel-plugin-transform-class-properties": "6.16.0",
-    "babel-plugin-transform-object-rest-spread": "6.16.0",
-    "babel-preset-es2015": "6.16.0",
+    "babel-plugin-transform-class-properties": "^6.22.0",
+    "babel-plugin-transform-object-rest-spread": "^6.22.0",
+    "babel-preset-es2015": "^6.22.0",
     "babel-preset-es2015-native-modules": "6.9.2",
-    "babel-preset-react": "6.16.0",
-    "coveralls": "2.11.12",
+    "babel-preset-react": "^6.22.0",
+    "coveralls": "2.11.15",
     "css-loader": "0.25.0",
-    "enzyme": "2.4.1",
-    "eslint": "3.6.1",
-    "eslint-config-airbnb": "12.0.0",
-    "eslint-plugin-import": "1.16.0",
-    "eslint-plugin-jsx-a11y": "2.2.2",
-    "eslint-plugin-react": "6.3.0",
+    "enzyme": "2.6.0",
+    "eslint": "3.11.1",
+    "eslint-config-airbnb": "13.0.0",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "2.2.3",
+    "eslint-plugin-react": "6.7.1",
     "extract-text-webpack-plugin": "2.0.0-beta.3",
+    "flow-bin": "^0.41.0",
+    "flow-typed": "^2.0.0",
     "immutable": "3.8.1",
-    "jest-cli": "15.1.1",
-    "node-sass": "3.10.0",
-    "react": "15.3.2",
-    "react-addons-test-utils": "15.3.2",
-    "react-dom": "15.3.2",
-    "react-redux": "4.4.5",
-    "react-test-renderer": "15.3.2",
+    "jest-cli": "18.1.0",
+    "node-sass": "3.13.0",
+    "prop-types": "^15.5.10",
+    "react": "15.4.1",
+    "react-addons-test-utils": "15.4.1",
+    "react-dom": "15.4.1",
+    "react-redux": "5.0.0",
+    "react-test-renderer": "15.4.1",
     "redux": "3.6.0",
     "redux-mock-store": "1.2.1",
     "redux-thunk": "2.1.0",
@@ -70,7 +74,7 @@
     "sass-loader": "4.0.2",
     "semantic-release": "4.3.5",
     "style-loader": "0.13.1",
-    "webpack": "2.1.0-beta.20"
+    "webpack": "2.1.0-beta.25"
   },
   "peerDependencies": {
     "immutable": "3",
@@ -83,13 +87,14 @@
   "dependencies": {
     "classnames": "2.2.5",
     "d3-drag": "1.0.0",
-    "d3-interpolate": "1.1.1",
-    "d3-scale": "1.0.2",
-    "d3-selection": "1.0.2",
+    "d3-interpolate": "1.1.2",
+    "d3-scale": "1.0.4",
+    "d3-selection": "1.0.3",
     "d3-shape": "1.0.3",
     "d3-zoom": "1.0.2",
-    "invariant": "2.2.1",
-    "lodash": "4.16.2",
+    "invariant": "2.2.2",
+    "lodash": "^4.17.5",
+    "prop-types": "^15.5.10",
     "react-immutable-proptypes": "2.1.0"
   }
 }
