@@ -3,8 +3,10 @@ import {
 	FLOWDESIGNER_LINK_SET_TARGET,
 	FLOWDESIGNER_LINK_SET_SOURCE,
 	FLOWDESIGNER_LINK_REMOVE,
-	FLOWDESIGNER_LINK_SET_ATTR,
-	FLOWDESIGNER_LINK_REMOVE_ATTR,
+	FLOWDESIGNER_LINK_SET_GRAPHICAL_ATTRIBUTES,
+	FLOWDESIGNER_LINK_REMOVE_GRAPHICAL_ATTRIBUTES,
+	FLOWDESIGNER_LINK_SET_DATA,
+	FLOWDESIGNER_LINK_REMOVE_DATA,
 } from '../constants/flowdesigner.constants';
 
 /**
@@ -15,13 +17,18 @@ import {
  * @param {string} linkType
  * @param {Object} attr
  */
-export const addLink = (linkId, sourceId, targetId, linkType, attr) => ({
+export const addLink = (
+	linkId,
+	sourceId,
+	targetId,
+	{ data = {}, graphicalAttributes = {} } = {},
+) => ({
 	type: FLOWDESIGNER_LINK_ADD,
 	linkId,
 	sourceId,
 	targetId,
-	linkType,
-	attr,
+	data,
+	graphicalAttributes,
 });
 
 /**
@@ -47,25 +54,47 @@ export const setLinkSource = (linkId, sourceId) => ({
 });
 
 /**
- * Ask to set attributes on link
+ * Ask to set graphical attributes on link
  * @param {string} linkId
  * @param {Object} attr
  */
-export const setLinkAttribute = (linkId, attr) => ({
-	type: FLOWDESIGNER_LINK_SET_ATTR,
+export const setLinkGraphicalAttributes = (linkId, graphicalAttributes) => ({
+	type: FLOWDESIGNER_LINK_SET_GRAPHICAL_ATTRIBUTES,
 	linkId,
-	attr,
+	graphicalAttributes,
 });
 
 /**
- * Ask to remove an attribute on target link
+ * Ask to remove an graphical attribute on target link
  * @param {string} linkId
  * @param {string} attrKey - the key of the attribute to be removed
  */
-export const removeLinkAttribute = (linkId, attrKey) => ({
-	type: FLOWDESIGNER_LINK_REMOVE_ATTR,
+export const removeLinkGraphicalAttribute = (linkId, graphicalAttributesKey) => ({
+	type: FLOWDESIGNER_LINK_REMOVE_GRAPHICAL_ATTRIBUTES,
 	linkId,
-	attrKey,
+	graphicalAttributesKey,
+});
+
+/**
+ * Ask to set data on link
+ * @param {string} linkId
+ * @param {Object} attr
+ */
+export const setLinkData = (linkId, data) => ({
+	type: FLOWDESIGNER_LINK_SET_DATA,
+	linkId,
+	data,
+});
+
+/**
+ * Ask to remove a data on target link
+ * @param {string} linkId
+ * @param {string} attrKey - the key of the attribute to be removed
+ */
+export const removeLinkData = (linkId, dataKey) => ({
+	type: FLOWDESIGNER_LINK_REMOVE_DATA,
+	linkId,
+	dataKey,
 });
 
 /**
@@ -73,7 +102,7 @@ export const removeLinkAttribute = (linkId, attrKey) => ({
  * @param {string} linkId
  * @return {Object}
  */
-export const removeLink = (linkId) => ({
+export const removeLink = linkId => ({
 	type: FLOWDESIGNER_LINK_REMOVE,
 	linkId,
 });
