@@ -8,121 +8,111 @@ import * as linkActions from './link.actions';
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
 
-describe('Check that link action creators generate proper' +
-	' action objects and perform checking', () => {
-	it('addLink', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_LINK_ADD',
-			linkId: 'linkId',
-			sourceId: 'sourceId',
-			targetId: 'targetId',
-			linkType: 'linkType',
-			attr: { selected: true },
-		}];
+describe(
+	'Check that link action creators generate proper' + ' action objects and perform checking',
+	() => {
+		it('addLink', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map(),
+					ports: new Map({
+						id1: { id: 'portId', portType: 'type' },
+						id2: { id: 'portId', portType: 'type' },
+					}),
+				},
+			});
 
-		const store = mockStore({
-			flowDesigner: {
-				links: new Map(),
-				ports: new Map({
-					id1: { id: 'portId', portType: 'type' },
-					id2: { id: 'portId', portType: 'type' },
+			store.dispatch(
+				linkActions.addLink('linkId', 'sourceId', 'targetId', 'linkType', {
+					graphicalAttributes: { selected: true },
 				}),
-			},
+			);
+			expect(store.getActions()).toMatchSnapshot();
 		});
 
-		store.dispatch(
-			linkActions.addLink('linkId', 'sourceId', 'targetId', 'linkType', { selected: true })
-		);
-		expect(store.getActions()).toEqual(expectedActions);
-	});
+		it('setLinkTarget', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map({ linkId: { id: 'linkId' } }),
+					ports: new Map({ id1: { id: 'portId', portType: 'type' } }),
+				},
+			});
 
-	it('setLinkTarget', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_LINK_SET_TARGET',
-			linkId: 'linkId',
-			targetId: 'portId',
-		}];
+			store.dispatch(linkActions.setLinkTarget('linkId', 'portId'));
 
-		const store = mockStore({
-			flowDesigner: {
-				links: new Map({ linkId: { id: 'linkId' } }),
-				ports: new Map({ id1: { id: 'portId', portType: 'type' } }),
-			},
+			expect(store.getActions()).toMatchSnapshot();
 		});
 
-		store.dispatch(linkActions.setLinkTarget('linkId', 'portId'));
+		it('setLinkSource', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map({ linkId: { id: 'linkId' } }),
+					ports: new Map({ id1: { id: 'portId', portType: 'type' } }),
+				},
+			});
 
-		expect(store.getActions()).toEqual(expectedActions);
-	});
+			store.dispatch(linkActions.setLinkSource('linkId', 'portId'));
 
-	it('setLinkSource', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_LINK_SET_SOURCE',
-			linkId: 'linkId',
-			sourceId: 'portId',
-		}];
-
-		const store = mockStore({
-			flowDesigner: {
-				links: new Map({ linkId: { id: 'linkId' } }),
-				ports: new Map({ id1: { id: 'portId', portType: 'type' } }),
-			},
+			expect(store.getActions()).toMatchSnapshot();
 		});
 
-		store.dispatch(linkActions.setLinkSource('linkId', 'portId'));
+		it('setLinkGraphicalAttributes', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map({ id: { id: 'linkId', linkType: 'type' } }),
+				},
+			});
 
-		expect(store.getActions()).toEqual(expectedActions);
-	});
+			store.dispatch(linkActions.setLinkGraphicalAttributes('id', { selected: true }));
 
-	it('setLinkAttribute', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_LINK_SET_ATTR',
-			linkId: 'id',
-			attr: { selected: true },
-		}];
-
-		const store = mockStore({
-			flowDesigner: {
-				links: new Map({ id: { id: 'linkId', linkType: 'type' } }),
-			},
+			expect(store.getActions()).toMatchSnapshot();
 		});
 
-		store.dispatch(linkActions.setLinkAttribute('id', { selected: true }));
+		it('removeLinkGrahicalAttribute', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map({ id: { id: 'linkId', linkType: 'type' } }),
+				},
+			});
 
-		expect(store.getActions()).toEqual(expectedActions);
-	});
+			store.dispatch(linkActions.removeLinkGraphicalAttribute('id', 'selected'));
 
-	it('removeLinkAttribute', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_LINK_REMOVE_ATTR',
-			linkId: 'id',
-			attrKey: 'selected',
-		}];
-
-		const store = mockStore({
-			flowDesigner: {
-				links: new Map({ id: { id: 'linkId', linkType: 'type' } }),
-			},
+			expect(store.getActions()).toMatchSnapshot();
 		});
 
-		store.dispatch(linkActions.removeLinkAttribute('id', 'selected'));
+		it('setLinkData', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map({ id: { id: 'linkId', linkType: 'type' } }),
+				},
+			});
 
-		expect(store.getActions()).toEqual(expectedActions);
-	});
+			store.dispatch(linkActions.setLinkData('id', { type: 'test' }));
 
-	it('removeLink', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_LINK_REMOVE',
-			linkId: 'id',
-		}];
-
-		const store = mockStore({
-			flowDesigner: {
-				links: new Map({ id: { id: 'linkId' } }),
-			},
+			expect(store.getActions()).toMatchSnapshot();
 		});
 
-		store.dispatch(linkActions.removeLink('id'));
-		expect(store.getActions()).toEqual(expectedActions);
-	});
-});
+		it('removeLinkData', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map({ id: { id: 'linkId', linkType: 'type' } }),
+				},
+			});
+
+			store.dispatch(linkActions.removeLinkData('id', 'type'));
+
+			expect(store.getActions()).toMatchSnapshot();
+		});
+
+		it('removeLink', () => {
+			const store = mockStore({
+				flowDesigner: {
+					links: new Map({ id: { id: 'linkId' } }),
+				},
+			});
+
+			store.dispatch(linkActions.removeLink('id'));
+			expect(store.getActions()).toMatchSnapshot();
+		});
+	},
+);
