@@ -1,13 +1,16 @@
 import {
+	FLOWDESIGNER_NODE_APPLY_MOVEMENT,
 	FLOWDESIGNER_NODE_MOVE,
 	FLOWDESIGNER_NODE_MOVE_END,
 	FLOWDESIGNER_NODE_ADD,
-	FLOWDESIGNER_NODE_SET_ATTR,
-	FLOWDESIGNER_NODE_REMOVE_ATTR,
+	FLOWDESIGNER_NODE_SET_TYPE,
+	FLOWDESIGNER_NODE_SET_GRAPHICAL_ATTRIBUTES,
+	FLOWDESIGNER_NODE_REMOVE_GRAPHICAL_ATTRIBUTES,
+	FLOWDESIGNER_NODE_SET_DATA,
+	FLOWDESIGNER_NODE_REMOVE_DATA,
 	FLOWDESIGNER_NODE_SET_SIZE,
 	FLOWDESIGNER_NODE_REMOVE,
 } from '../constants/flowdesigner.constants';
-
 
 /**
  * Ask for node creation and injection into current dataflow
@@ -18,13 +21,12 @@ import {
  * @param {Object} attr
  * @return {Object}
  */
-export const addNode = (nodeId, nodePosition, nodeSize, nodeType, attr) => ({
+export const addNode = (nodeId, nodeType, { data = {}, graphicalAttributes = {} } = {}) => ({
 	type: FLOWDESIGNER_NODE_ADD,
 	nodeId,
-	nodePosition,
-	nodeSize,
 	nodeType,
-	attr,
+	data,
+	graphicalAttributes,
 });
 
 /**
@@ -33,10 +35,25 @@ export const addNode = (nodeId, nodePosition, nodeSize, nodeType, attr) => ({
  * @param {{x: number, y: number}} nodePosition - the new absolute position of the node
  * @return {Object}
  */
-export const moveNodeTo = (nodeId, nodePosition) => ({
-	type: FLOWDESIGNER_NODE_MOVE,
-	nodeId,
-	nodePosition,
+export function moveNodeTo(nodeId, nodePosition) {
+	return {
+		type: FLOWDESIGNER_NODE_MOVE,
+		nodeId,
+		nodePosition,
+	};
+}
+
+/**
+ * Ask to apply the same movement to multiples nodesId
+ * @param nodesId {array<string>} list of nodeId
+ * @param movement {Object} relative movement to apply on all nodes
+ *
+ * @return {Object}
+ */
+export const applyMovementTo = (nodesId, movement) => ({
+	type: FLOWDESIGNER_NODE_APPLY_MOVEMENT,
+	nodesId,
+	movement,
 });
 
 /**
@@ -45,11 +62,13 @@ export const moveNodeTo = (nodeId, nodePosition) => ({
  * @param {{x: number, y: number}} nodePosition - the new absolute position of the node
  * @return {Object}
  */
-export const moveNodeToEnd = (nodeId, nodePosition) => ({
-	type: FLOWDESIGNER_NODE_MOVE_END,
-	nodeId,
-	nodePosition,
-});
+export function moveNodeToEnd(nodeId, nodePosition) {
+	return {
+		type: FLOWDESIGNER_NODE_MOVE_END,
+		nodeId,
+		nodePosition,
+	};
+}
 
 /**
  * set node size
@@ -64,25 +83,61 @@ export const setNodeSize = (nodeId, nodeSize) => ({
 });
 
 /**
- * Give the ability to set any data onto the node
+ * Ask for node creation and injection into current dataflow
  * @param {string} nodeId
- * @param {Object} attr
+ * @param {string} nodeType
+ * @return {Object}
  */
-export const setNodeAttribute = (nodeId, attr) => ({
-	type: FLOWDESIGNER_NODE_SET_ATTR,
+export function setNodeType(nodeId, nodeType) {
+	return {
+		type: FLOWDESIGNER_NODE_SET_TYPE,
+		nodeId,
+		nodeType,
+	};
+}
+
+/**
+ * Give the ability to a graphical attribute onto the node
+ * @param {string} nodeId
+ * @param {Object} graphicalAttributes
+ */
+export const setNodeGraphicalAttributes = (nodeId, graphicalAttributes) => ({
+	type: FLOWDESIGNER_NODE_SET_GRAPHICAL_ATTRIBUTES,
 	nodeId,
-	attr,
+	graphicalAttributes,
 });
 
 /**
- * Ask to remove an attribute on target node
+ * Ask to remove a graphical attribute on target node
  * @param {string} nodeId
- * @param {string} attrKey - the key of the attribute to be removed
+ * @param {string} graphicalAttributesKey - the key of the attribute to be removed
  */
-export const removeNodeAttribute = (nodeId, attrKey) => ({
-	type: FLOWDESIGNER_NODE_REMOVE_ATTR,
+export const removeNodeGraphicalAttribute = (nodeId, graphicalAttributesKey) => ({
+	type: FLOWDESIGNER_NODE_REMOVE_GRAPHICAL_ATTRIBUTES,
 	nodeId,
-	attrKey,
+	graphicalAttributesKey,
+});
+
+/**
+ * Give the ability to set data onto a node
+ * @param {string} nodeId
+ * @param {Object} data
+ */
+export const setNodeData = (nodeId, data) => ({
+	type: FLOWDESIGNER_NODE_SET_DATA,
+	nodeId,
+	data,
+});
+
+/**
+ * Ask to remove a graphical attribute on target node
+ * @param {string} nodeId
+ * @param {string} dataKey - the key of the data to be removed
+ */
+export const removeNodeData = (nodeId, dataKey) => ({
+	type: FLOWDESIGNER_NODE_REMOVE_DATA,
+	nodeId,
+	dataKey,
 });
 
 /**
