@@ -4,23 +4,13 @@ import thunk from 'redux-thunk';
 import { Map, OrderedMap } from 'immutable';
 
 import * as nodeActions from './node.actions';
-
+import { FLOWDESIGNER_NODE_SET_TYPE } from '../constants/flowdesigner.constants';
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
 
-describe('Check that node action creators generate proper' +
-	' action objects and perform checking', () => {
+describe('Check that node action creators generate proper action objects and perform checking', () => {
 	it('addNode generate action with 0 configuration', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_NODE_ADD',
-			nodeId: 'id',
-			nodePosition: { x: 75, y: 75 },
-			nodeSize: { width: 50, heigth: 50 },
-			nodeType: 'nodeType',
-			attr: {},
-		}];
-
 		const store = mockStore({
 			flowDesigner: {
 				nodes: new Map({}),
@@ -28,22 +18,23 @@ describe('Check that node action creators generate proper' +
 		});
 
 		store.dispatch(
-			nodeActions.addNode('id', { x: 75, y: 75 }, { width: 50, heigth: 50 }, 'nodeType', {})
+			nodeActions.addNode('id', 'label', 'description', {
+				data: {},
+				graphicalAttributes: {
+					nodePosition: { x: 75, y: 75 },
+					nodeSize: { width: 50, heigth: 50 },
+					nodeType: 'nodeType',
+				},
+			}),
 		);
 
-		expect(store.getActions()).toEqual(expectedActions);
+		expect(store.getActions()).toMatchSnapshot();
 	});
 
 	it('moveNode generate a proper action object witch nodeId and nodePosition parameter', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_NODE_MOVE',
-			nodeId: 'nodeId',
-			nodePosition: { x: 10, y: 20 },
-		}];
-
 		const store = mockStore({
 			flowDesigner: {
-				nodes: new Map({ nodeId: { id: 'nodeId', nodeType: 'type' } }),
+				nodes: new Map({ nodeId: { id: 'nodeId', type: 'type' } }),
 				nodeTypes: new Map({
 					type: new Map({
 						component: { calculatePortPosition: () => ({}) },
@@ -55,76 +46,102 @@ describe('Check that node action creators generate proper' +
 
 		store.dispatch(nodeActions.moveNodeTo('nodeId', { x: 10, y: 20 }, {}));
 
-		expect(store.getActions()).toEqual(expectedActions);
+		expect(store.getActions()).toMatchSnapshot();
 	});
 
 	it('setNodeSize', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_NODE_SET_SIZE',
-			nodeId: 'nodeId',
-			nodeSize: { width: 100, height: 100 },
-		}];
-
 		const store = mockStore({
 			flowDesigner: {
-				nodes: new Map({ nodeId: { id: 'nodeId', nodeType: 'type' } }),
+				nodes: new Map({ nodeId: { id: 'nodeId', type: 'type' } }),
 			},
 		});
 
 		store.dispatch(nodeActions.setNodeSize('nodeId', { width: 100, height: 100 }));
-		expect(store.getActions()).toEqual(expectedActions);
+		expect(store.getActions()).toMatchSnapshot();
 	});
 
-	it('setNodeAttribute', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_NODE_SET_ATTR',
-			nodeId: 'id',
-			attr: { selected: true },
-		}];
-
+	it('setNodeType', () => {
+		const nodeId = 'nodeId';
+		const nodeType = 'newNodeType';
 		const store = mockStore({
 			flowDesigner: {
-				nodes: new Map({ id: { id: 'nodeId', nodeType: 'type' } }),
+				nodes: new Map({ nodeId: { id: nodeId, type: 'type' } }),
 			},
 		});
 
-		store.dispatch(nodeActions.setNodeAttribute('id', { selected: true }));
-
-		expect(store.getActions()).toEqual(expectedActions);
+		store.dispatch(nodeActions.setNodeType(nodeId, nodeType));
+		expect(store.getActions()[0]).toEqual({
+			type: FLOWDESIGNER_NODE_SET_TYPE,
+			nodeId,
+			nodeType,
+		});
 	});
 
-	it('removeNodeAttribute', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_NODE_REMOVE_ATTR',
-			nodeId: 'id',
-			attrKey: 'selected',
-		}];
-
+	it('setNodeGraphicalAttributes', () => {
 		const store = mockStore({
 			flowDesigner: {
-				nodes: new Map({ id: { id: 'nodeId', nodeType: 'type' } }),
+				nodes: new Map({ id: { id: 'nodeId', type: 'type' } }),
 			},
 		});
 
-		store.dispatch(nodeActions.removeNodeAttribute('id', 'selected'));
+		store.dispatch(nodeActions.setNodeGraphicalAttributes('id', { selected: true }));
 
-		expect(store.getActions()).toEqual(expectedActions);
+		expect(store.getActions()).toMatchSnapshot();
+	});
+
+	it('removeNodeGraphicalAttribute', () => {
+		const store = mockStore({
+			flowDesigner: {
+				nodes: new Map({ id: { id: 'nodeId', type: 'type' } }),
+			},
+		});
+
+		store.dispatch(nodeActions.removeNodeGraphicalAttribute('id', 'selected'));
+
+		expect(store.getActions()).toMatchSnapshot();
+	});
+
+	it('setNodeData', () => {
+		const store = mockStore({
+			flowDesigner: {
+				nodes: new Map({ id: { id: 'nodeId', type: 'type' } }),
+			},
+		});
+
+		store.dispatch(nodeActions.setNodeData('id', { type: 'test' }));
+
+		expect(store.getActions()).toMatchSnapshot();
+	});
+
+	it('removeNodeData', () => {
+		const store = mockStore({
+			flowDesigner: {
+				nodes: new Map({
+					id: { id: 'nodeId', type: 'type', data: new Map({ testProperties: 'testProperties' }) },
+				}),
+			},
+		});
+
+		store.dispatch(nodeActions.removeNodeData('id', 'type'));
+
+		expect(store.getActions()).toMatchSnapshot();
 	});
 
 	it('removeNode', () => {
-		const expectedActions = [{
-			type: 'FLOWDESIGNER_NODE_REMOVE',
-			nodeId: 'id',
-		}];
-
 		const store = mockStore({
 			flowDesigner: {
-				nodes: new Map({ id: { id: 'nodeId', nodeType: 'type' } }),
+				nodes: new Map({ id: { id: 'nodeId', type: 'type' } }),
 			},
 		});
 
 		store.dispatch(nodeActions.removeNode('id'));
 
-		expect(store.getActions()).toEqual(expectedActions);
+		expect(store.getActions()).toMatchSnapshot();
+	});
+});
+
+describe('applyMovementTo', () => {
+	it('generate proper action', () => {
+		expect(nodeActions.applyMovementTo([1, 2, 3], { x: 10, y: 5 })).toMatchSnapshot();
 	});
 });
