@@ -1,58 +1,105 @@
-import {
-	FLOWDESIGNER_PORT_ADD,
-	FLOWDESIGNER_PORT_ADDS,
-	FLOWDESIGNER_PORT_SET_ATTR,
-	FLOWDESIGNER_PORT_REMOVE_ATTR,
-	FLOWDESIGNER_PORT_REMOVE,
-} from '../constants/flowdesigner.constants';
+/* @flow */
+
+import type {
+	PortGraphicalAttributes,
+	Port,
+	PortData,
+	PortAction,
+	PortActionAdd,
+	Id,
+} from '../flow-typed';
+
 /**
  * return an action to create a new port
  * @param {string} nodeId - identifier of the node to wich the created connector should be attached
- * @param {string} portId
+ * @param {string} id
  * @param {string} portType
- * @param {Object} attr
+ * @param {Object} attributes
  */
-export const addPort = (nodeId, portId, portType, attr) => ({
-	type: FLOWDESIGNER_PORT_ADD,
-	nodeId,
-	portId,
-	portType,
-	attr,
-});
+export function addPort(
+	nodeId: Id,
+	id: Id,
+	{ data, graphicalAttributes }: { data: PortData, graphicalAttributes: PortGraphicalAttributes },
+): PortActionAdd {
+	return {
+		type: 'FLOWDESIGNER_PORT_ADD',
+		nodeId,
+		id,
+		data,
+		graphicalAttributes,
+	};
+}
 
-export const addPorts = (nodeId, ports) => ({
-	type: FLOWDESIGNER_PORT_ADDS,
-	nodeId,
-	ports,
-});
+export function addPorts(nodeId: Id, ports: Array<Port>): PortAction {
+	return {
+		type: 'FLOWDESIGNER_PORT_ADDS',
+		nodeId,
+		ports,
+	};
+}
 
 /**
  * return an action to set port attributes
  * @param {string} portId
- * @param {Object} attr
+ * @param {Object} graphicalAttributes
  */
-export const setPortAttribute = (portId, attr) => ({
-	type: FLOWDESIGNER_PORT_SET_ATTR,
-	portId,
-	attr,
-});
+export function setPortGraphicalAttribute(portId: Id, graphicalAttributes: {}): PortAction {
+	return {
+		type: 'FLOWDESIGNER_PORT_SET_GRAPHICAL_ATTRIBUTES',
+		portId,
+		graphicalAttributes,
+	};
+}
 
 /**
  * Ask to remove an attribute on target port
  * @param {string} portId
- * @param {string} attrKey - the key of the attribute to be removed
+ * @param {string} graphicalAttributesKey - the key of the attribute to be removed
  */
-export const removePortAttribute = (portId, attrKey) => ({
-	type: FLOWDESIGNER_PORT_REMOVE_ATTR,
-	portId,
-	attrKey,
-});
+export function removePortGraphicalAttribute(
+	portId: Id,
+	graphicalAttributesKey: string,
+): PortAction {
+	return {
+		type: 'FLOWDESIGNER_PORT_REMOVE_GRAPHICAL_ATTRIBUTES',
+		portId,
+		graphicalAttributesKey,
+	};
+}
+
+/**
+ * return an action to set port attributes
+ * @param {string} portId
+ * @param {Object} graphicalAttributes
+ */
+export function setPortdata(portId: Id, data: Object): PortAction {
+	return {
+		type: 'FLOWDESIGNER_PORT_SET_DATA',
+		portId,
+		data,
+	};
+}
+
+/**
+ * Ask to remove an attribute on target port
+ * @param {string} portId
+ * @param {string} datasKey - the key of the attribute to be removed
+ */
+export function removePortData(portId: Id, dataKey: string): PortAction {
+	return {
+		type: 'FLOWDESIGNER_PORT_REMOVE_DATA',
+		portId,
+		dataKey,
+	};
+}
 
 /**
  * return an action to remove port and all attached links
  * @param {string} portId
  */
-export const removePort = portId => ({
-	type: FLOWDESIGNER_PORT_REMOVE,
-	portId,
-});
+export function removePort(portId: Id): PortAction {
+	return {
+		type: 'FLOWDESIGNER_PORT_REMOVE',
+		portId,
+	};
+}
