@@ -1,27 +1,38 @@
-jest.mock('d3-selection', () => ({
-    select() {
-        return { data() {}, call() {} };
-    },
-}));
-
 import React from 'react';
 import renderer from 'react-test-renderer';
 
 import AbstractNode from './AbstractNode.component';
-import { NodeRecord, PositionRecord } from '../../constants/flowdesigner.model';
+import {
+	NodeGraphicalAttributes,
+	NodeRecord,
+	PositionRecord,
+	SizeRecord,
+} from '../../constants/flowdesigner.model';
 
+jest.mock('d3-selection', () => ({
+	select() {
+		return { data() {}, call() {} };
+	},
+}));
 
-describe('<AbstractNode /> renders correctly', () => {
-    it('<AbstractNode /> renders correctly', () => {
-        const node = new NodeRecord({
-            id: 'nodeId',
-            position: new PositionRecord({ x: 100, y: 100 }),
-        });
-        const tree = renderer.create(
-          <AbstractNode node={node}>
-            <rect />
-          </AbstractNode>
-        ).toJSON();
-        expect(tree).toMatchSnapshot();
-    });
+const noOp = () => {};
+
+describe('<AbstractNode />', () => {
+	it('renders correctly', () => {
+		const node = new NodeRecord({
+			id: 'id',
+			graphicalAttributes: new NodeGraphicalAttributes({
+				position: new PositionRecord({ x: 100, y: 100 }),
+				nodeSize: new SizeRecord({ width: 125, height: 75 }),
+			}),
+		});
+		const tree = renderer
+			.create(
+				<AbstractNode node={node} moveNodeTo={noOp} moveNodeToEnd={noOp}>
+					<rect />
+				</AbstractNode>,
+			)
+			.toJSON();
+		expect(tree).toMatchSnapshot();
+	});
 });
