@@ -1,30 +1,42 @@
-import ReactTestUtils from 'react-addons-test-utils';
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import { Map } from 'immutable';
 
-import { NodeRecord, PositionRecord, SizeRecord } from '../../constants/flowdesigner.model';
-import AbstractNode from './AbstractNode.component.jsx';
+import {
+	NodeGraphicalAttributes,
+	NodeRecord,
+	PositionRecord,
+	SizeRecord,
+} from '../../constants/flowdesigner.model';
+import AbstractNode, { ABSTRACT_NODE_INVARIANT } from './AbstractNode.component';
 
 const node = new NodeRecord({
 	id: 'id',
-	position: new PositionRecord({ x: 100, y: 50 }),
-	nodeSize: new SizeRecord({ width: 125, height: 75 }),
-	attr: new Map(),
+	graphicalAttributes: new NodeGraphicalAttributes({
+		position: new PositionRecord({ x: 100, y: 50 }),
+		nodeSize: new SizeRecord({ width: 125, height: 75 }),
+	}),
 });
 
+function noOp() {}
 
 describe('Testing <AbstractNode>', () => {
-
 	it('should create a bare node component with provided position', () => {
-		const wrapper = shallow(<AbstractNode node={node}><rect /></AbstractNode>);
+		const wrapper = shallow(
+			<AbstractNode node={node} moveNodeTo={noOp} moveNodeToEnd={noOp}>
+				<rect />
+			</AbstractNode>,
+		);
 		const rect = wrapper.find('g[transform]');
-		expect(rect.prop('transform')).toBe('translate(100,50)');
+		expect(rect.prop('transform')).toBe('translate(100, 50)');
 	});
 
 	it('call the injected onClick action when clicked', () => {
 		const onClick = jasmine.createSpy('onClick');
-		const wrapper = shallow(<AbstractNode node={node} onClick={onClick}><rect /></AbstractNode>);
+		const wrapper = shallow(
+			<AbstractNode node={node} onClick={onClick} moveNodeTo={noOp} moveNodeToEnd={noOp}>
+				<rect />
+			</AbstractNode>,
+		);
 		wrapper.find('g[transform]').simulate('click');
 		expect(onClick.and.identity()).toEqual('onClick');
 		expect(onClick).toHaveBeenCalled();
@@ -33,24 +45,23 @@ describe('Testing <AbstractNode>', () => {
 
 	// if anyone got a clue on how to test react + d3 events
 
-	xit('call the injected onDragStart action when drag action start', (done) => {
+	xit('call the injected onDragStart action when drag action start', done => {
 		const evt = document.createEvent('HTMLEvents');
 		evt.initEvent('click', false, true);
 		const onDragStart = jest.fn();
 		mount(
-			<AbstractNode node={node} onClick={onDragStart}>
+			<AbstractNode node={node} onClick={onDragStart} moveNodeTo={noOp} moveNodeToEnd={noOp}>
 				<rect />
-			</AbstractNode>
-			, { attachTo: document.body }
+			</AbstractNode>,
+			{ attachTo: document.body },
 		);
 
-		document.querySelector('g g').addEventListener('click', function() {
-			console.error('SUPER');
+		document.querySelector('g g').addEventListener('click', () => {
 			done();
 		});
 		document.querySelector('g g').dispatchEvent(new window.MouseEvent('click'));
 		expect(onDragStart.mock.calls.length).toBeNull(1);
-		fail()
+		fail();
 	});
 
 	xit('call the injected onDrag action when drag action start', () => {
@@ -63,10 +74,7 @@ describe('Testing <AbstractNode>', () => {
 
 	it('should fire an error if its rendered without a children set up', () => {
 		expect(() => {
-			shallow(<AbstractNode node={node} />);
-		}).toThrowError(
-			'<AbstractNode /> should not be used without giving it a children' +
-				'ex: <AbstractNode><rect /></AbstractNode>'
-		);
+			shallow(<AbstractNode node={node} moveNodeTo={noOp} moveNodeToEnd={noOp} />);
+		}).toThrowError(ABSTRACT_NODE_INVARIANT);
 	});
 });
