@@ -3,26 +3,38 @@ import renderer from 'react-test-renderer';
 import { Map } from 'immutable';
 
 import NodesRenderer from './NodesRenderer.component';
-import { NodeRecord } from '../../constants/flowdesigner.model';
+import { NodeRecord, NodeGraphicalAttributes } from '../../constants/flowdesigner.model';
 
-const MockNode = () => (
-  <span>MockLink</span>
-);
+const MockNode = () => <span>MockNodes</span>;
 
+const noOp = () => {};
 
-describe('<LinksRenderer /> renders correctly', () => {
-    it('<LinksRenderer /> renders correctly', () => {
-        const nodes = new Map().set('id', new NodeRecord({
-            id: 'id',
-            nodeType: 'id',
-
-        }));
-        const nodeTypeMap = {
-            id: { id: 'id', component: MockNode },
-        };
-        const tree = renderer.create(
-          <NodesRenderer nodes={nodes} nodeTypeMap={nodeTypeMap} />
-        ).toJSON();
-        expect(tree).toMatchSnapshot();
-    });
+describe('<NodesRenderer />', () => {
+	it('renders correctly', () => {
+		const nodes = new Map().set(
+			'id',
+			new NodeRecord({
+				id: 'id',
+				type: 'id',
+				graphicalAttributes: new NodeGraphicalAttributes({
+					nodeType: 'id',
+				}),
+			}),
+		);
+		const nodeTypeMap = {
+			id: { id: 'id', component: MockNode },
+		};
+		const tree = renderer
+			.create(
+				<NodesRenderer
+					nodes={nodes}
+					nodeTypeMap={nodeTypeMap}
+					moveNodeTo={noOp}
+					moveNodeToEnd={noOp}
+					snapToGrid
+				/>,
+			)
+			.toJSON();
+		expect(tree).toMatchSnapshot();
+	});
 });
