@@ -5,22 +5,26 @@ import { Map } from 'immutable';
 import PortsRenderer from './PortsRenderer.component';
 import { PortRecord } from '../../constants/flowdesigner.model';
 
-const MockPort = () => (
-  <span>MockPort</span>
-);
+const MockPort = () => <span>MockPort</span>;
 
-describe('<PortsRenderer /> renders correctly', () => {
-    it('<PortsRenderer /> renders correctly', () => {
-        const ports = new Map().set('id', new PortRecord({
-            id: 'id',
-			portType: 'id',
-        }));
+describe('<PortsRenderer />', () => {
+	it('renders correctly', () => {
+		const ports = new Map().set(
+			'id',
+			new PortRecord({
+				id: 'id',
+				nodeId: 'nodeId',
+				graphicalAttributes: new Map({
+					portType: 'id',
+				}),
+			}),
+		);
 		const portTypeMap = {
-            id: { id: 'id', component: MockPort },
-        };
-        const tree = renderer.create(
-          <PortsRenderer ports={ports} portTypeMap={portTypeMap} />
-        ).toJSON();
-        expect(tree).toMatchSnapshot();
-    });
+			id: { id: 'id', component: MockPort },
+		};
+		const tree = renderer
+			.create(<PortsRenderer ports={ports} portTypeMap={portTypeMap} />)
+			.toJSON();
+		expect(tree).toMatchSnapshot();
+	});
 });
