@@ -1,235 +1,159 @@
-import { Map, OrderedMap } from 'immutable';
+import { Map } from 'immutable';
 
-import { reducer, calculatePortsPosition } from './flow.reducer';
+import { reducer, calculatePortsPosition, defaultState } from './flow.reducer';
 import * as nodeActions from '../actions/node.actions';
 import * as portActions from '../actions/port.actions';
-import {
-	NodeRecord,
-	PortRecord,
-	SizeRecord,
-	PositionRecord,
-} from '../constants/flowdesigner.model';
+import { NodeRecord, PortRecord } from '../constants/flowdesigner.model';
+import { PORT_SOURCE } from '../constants/flowdesigner.constants';
 
-describe('FLOWDESIGNER_FLOW_ADD_ELEMENTS is batching elements creation', () => {
+describe('FLOWDESIGNER_FLOW_ADD_ELEMENTS', () => {
 	it('should batch one element creation', () => {
-		expect(reducer(new Map(), {
-			type: 'FLOWDESIGNER.FLOW_ADD_ELEMENTS',
-			listOfActionCreation: [
-				nodeActions.addNode(
-					'nodeId',
-					{ x: 10, y: 10 },
-					{ height: 10, width: 10 },
-					undefined,
-					{}
-				),
-			],
-		})).toEqual(new Map()
-			.set('nodes', new Map()
-				.set('nodeId', new NodeRecord({
-					id: 'nodeId',
-					position: new PositionRecord({
-						x: 10,
-						y: 10,
+		expect(
+			reducer(defaultState, {
+				type: 'FLOWDESIGNER.FLOW_ADD_ELEMENTS',
+				listOfActionCreation: [
+					nodeActions.addNode('nodeId', undefined, {
+						data: {},
+						graphicalAttributes: {
+							nodeSize: { height: 10, width: 10 },
+							position: { x: 10, y: 10 },
+						},
 					}),
-					nodeSize: new SizeRecord({
-						height: 10,
-						width: 10,
-					}),
-					nodeType: undefined,
-					attr: new Map(),
-				}))
-			));
+				],
+			}),
+		).toMatchSnapshot();
 	});
 
 	it('should batch many elements creation', () => {
-		expect(reducer(new Map(), {
-			type: 'FLOWDESIGNER.FLOW_ADD_ELEMENTS',
-			listOfActionCreation: [
-				nodeActions.addNode(
-					'nodeId',
-					{ x: 10, y: 10 },
-					{ height: 10, width: 10 },
-					undefined,
-					{}
-				),
-				nodeActions.addNode(
-					'node2',
-					{ x: 10, y: 10 },
-					{ height: 10, width: 10 },
-					undefined,
-					{}
-				),
-				portActions.addPort(
-					'nodeId',
-					'portId',
-					undefined,
-					{},
-				),
-			],
-		})).toEqual(new Map()
-			.set('nodes', new Map()
-				.set('nodeId', new NodeRecord({
-					id: 'nodeId',
-					position: new PositionRecord({
-						x: 10,
-						y: 10,
+		expect(
+			reducer(defaultState, {
+				type: 'FLOWDESIGNER.FLOW_ADD_ELEMENTS',
+				listOfActionCreation: [
+					nodeActions.addNode('nodeId', undefined, {
+						data: {},
+						graphicalAttributes: {
+							nodeSize: { height: 10, width: 10 },
+							position: { x: 10, y: 10 },
+						},
 					}),
-					nodeSize: new SizeRecord({
-						height: 10,
-						width: 10,
+					nodeActions.addNode('node2', undefined, {
+						data: {},
+						graphicalAttributes: {
+							nodeSize: { height: 10, width: 10 },
+							position: { x: 10, y: 10 },
+						},
 					}),
-					nodeType: undefined,
-					attr: new Map(),
-				}))
-				.set('node2', new NodeRecord({
-					id: 'node2',
-					position: new PositionRecord({
-						x: 10,
-						y: 10,
+					portActions.addPort('nodeId', 'portId', {
+						graphicalAttributes: {
+							properties: {
+								type: PORT_SOURCE,
+							},
+						},
 					}),
-					nodeSize: new SizeRecord({
-						height: 10,
-						width: 10,
-					}),
-					nodeType: undefined,
-					attr: new Map(),
-				}))
-			).set('ports', new Map()
-				.set('portId', new PortRecord({
-					id: 'portId',
-					nodeId: 'nodeId',
-					portType: undefined,
-					position: undefined,
-					attr: new Map(),
-				})))
-			);
+				],
+			}),
+		).toMatchSnapshot();
 	});
 
 	it('should handle throwing sub reducer by returning old state', () => {
-		expect(reducer(new Map(), {
-			type: 'FLOWDESIGNER.FLOW_ADD_ELEMENTS',
-			listOfActionCreation: [
-				nodeActions.addNode(
-					'nodeId',
-					{ x: 10, y: 10 },
-					{ height: 10, width: 10 },
-					undefined,
-					{}
-				),
-				nodeActions.addNode(
-					'node2',
-					{ x: 10, y: 10 },
-					{ height: 10, width: 10 },
-					undefined,
-					{}
-				),
-				portActions.addPort(
-					'node3',
-					'portId',
-					undefined,
-					{},
-				),
-			],
-		})).toEqual(new Map());
+		expect(
+			reducer(defaultState, {
+				type: 'FLOWDESIGNER.FLOW_ADD_ELEMENTS',
+				listOfActionCreation: [
+					nodeActions.addNode('nodeId', undefined, {
+						data: {},
+						graphicalAttributes: {
+							nodeSize: { height: 10, width: 10 },
+							position: { x: 10, y: 10 },
+						},
+					}),
+					nodeActions.addNode('node2', undefined, {
+						data: {},
+						graphicalAttributes: {
+							nodeSize: { height: 10, width: 10 },
+							position: { x: 10, y: 10 },
+						},
+					}),
+					portActions.addPort('node3', 'portId', {
+						data: undefined,
+						graphicalAttributes: undefined,
+					}),
+				],
+			}),
+		).toMatchSnapshot();
 	});
 });
 
 describe('FLOWDESIGNER_FLOW_LOAD should reset old flow state and load news not touching flow config', () => {
 	it('should load elements', () => {
-		expect(reducer(new Map({ nports: new Map(), links: new Map(), ports: new Map(), nodeTypes: new Map() }), {
-			type: 'FLOWDESIGNER.FLOW_LOAD',
-			listOfActionCreation: [
-				nodeActions.addNode(
-					'nodeId',
-					{ x: 10, y: 10 },
-					{ height: 10, width: 10 },
-					undefined,
-					{}
-				),
-				nodeActions.addNode(
-					'node2',
-					{ x: 10, y: 10 },
-					{ height: 10, width: 10 },
-					undefined,
-					{}
-				),
-				portActions.addPort(
-					'nodeId',
-					'portId',
-					undefined,
-					{},
-				),
-			],
-		})).toEqual(new Map()
-			.set('nodes', new Map()
-				.set('nodeId', new NodeRecord({
-					id: 'nodeId',
-					position: new PositionRecord({
-						x: 10,
-						y: 10,
+		expect(
+			reducer(defaultState, {
+				type: 'FLOWDESIGNER.FLOW_LOAD',
+				listOfActionCreation: [
+					nodeActions.addNode('nodeId', undefined, {
+						data: {},
+						graphicalAttributes: {
+							nodeSize: { height: 10, width: 10 },
+							position: { x: 10, y: 10 },
+						},
 					}),
-					nodeSize: new SizeRecord({
-						height: 10,
-						width: 10,
+					nodeActions.addNode('node2', undefined, {
+						data: {},
+						graphicalAttributes: {
+							nodeSize: { height: 10, width: 10 },
+							position: { x: 10, y: 10 },
+						},
 					}),
-					nodeType: undefined,
-					attr: new Map(),
-				}))
-				.set('node2', new NodeRecord({
-					id: 'node2',
-					position: new PositionRecord({
-						x: 10,
-						y: 10,
+					portActions.addPort('nodeId', 'portId', {
+						graphicalAttributes: {
+							properties: {
+								type: PORT_SOURCE,
+							},
+						},
 					}),
-					nodeSize: new SizeRecord({
-						height: 10,
-						width: 10,
-					}),
-					nodeType: undefined,
-					attr: new Map(),
-				}))
-			).set('links', new Map())
-			.set('ports', new OrderedMap()
-				.set('portId', new PortRecord({
-					id: 'portId',
-					nodeId: 'nodeId',
-					portType: undefined,
-					position: undefined,
-					attr: new Map(),
-				}))
-			)
-			.set('nodeTypes', new Map())
-			.set('transform', {
-				k: 1,
-				x: 0,
-				y: 0,
-			})
-		);
+				],
+			}),
+		).toMatchSnapshot();
 	});
 });
 
+describe('FLOWDESIGNER_PAN_TO set a calculated transformation into transformToApply', () => {
+	it('', () => {
+		expect(
+			reducer(defaultState, {
+				type: 'FLOWDESIGNER_PAN_TO',
+				x: 400,
+				y: 400,
+			}),
+		).toMatchSnapshot();
+	});
+});
 
 describe('calculatePortsPosition behavior', () => {
-	const state = new Map()
-		.set('nodes', new Map()
-			.set('42', new NodeRecord({
-				id: '42',
-				position: new PositionRecord({}),
-				nodeSize: new SizeRecord({}),
-				nodeType: '42',
-			}))
+	const state = defaultState
+		.set(
+			'nodes',
+			new Map().set(
+				'42',
+				new NodeRecord({
+					id: '42',
+					graphicalAttributes: new Map({
+						nodeType: '42',
+					}),
+				}),
+			),
 		)
-		.set('ports', new Map()
-			.set('42', new PortRecord({
-				id: '42',
-				nodeId: '42',
-			}))
+		.set(
+			'ports',
+			new Map().set(
+				'42',
+				new PortRecord({
+					id: '42',
+					nodeId: '42',
+				}),
+			),
 		)
-		.set('nodeTypes', new Map()
-			.set('42', new Map()
-				.set('component', {}),
-			)
-		);
+		.set('nodeTypes', new Map().set('42', new Map().set('component', {})));
 
 	it('should trigger only if NODE/PORT/FLOW action are dispatched', () => {
 		const calculatePortPosition = jest.fn();
@@ -247,7 +171,7 @@ describe('calculatePortsPosition behavior', () => {
 	});
 
 	it('should not trigger on FLOWDESIGNER_NODE_REMOVE and FLOWDESIGNER_PORT_REMOVE', () => {
-				const calculatePortPosition = jest.fn();
+		const calculatePortPosition = jest.fn();
 		const givenState = state.setIn(['nodeTypes', '42', 'component'], { calculatePortPosition });
 		calculatePortsPosition(givenState, {
 			type: 'FLOWDESIGNER_NODE_REMOVE',
